backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0

# ── Database ───────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ── Redis / Caching ────────────────────────────────────────
redis[asyncio]==5.0.4
aioredis==2.0.1

# ── Task Queue ─────────────────────────────────────────────
celery[redis]==5.4.0
flower==2.0.1

# ── Auth & Security ────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ── HTTP / Scraping ────────────────────────────────────────
httpx[http2]==0.27.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.2

# ── Media / Download ───────────────────────────────────────
yt-dlp==2024.5.27
Pillow==10.3.0

# ── Rate Limiting ──────────────────────────────────────────
slowapi==0.1.9

# ── Config & Validation ────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.1
python-dotenv==1.0.1

# ── Logging ────────────────────────────────────────────────
structlog==24.2.0

# ── Utilities ──────────────────────────────────────────────
python-magic==0.4.27
aiofiles==23.2.1
click==8.1.7
